requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pymupdf>=1.24.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
torch>=2.0.0
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
ragas>=0.2.0
langfuse>=2.0.0
streamlit>=1.40.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pyyaml>=6.0.0
httpx>=0.27.0
tqdm>=4.66.0
